refactor: extract keyboard shortcuts into modules/shortcuts.js
Moves SECTION 22 out of app.js, taking it from 12,363 to 11,914 lines.

setupKeyboardShortcuts owns the two chord flags that nothing else reads
(tabChordUsed, tabPressedInTerminal). tabHeld stays in app.js because the
wheel handlers use it for Tab+scroll canvas panning, so the module reads and
writes it through the injected context, along with the HUD visibility flags
and the double-tap timestamp.

Verified against a running instance: Tab+A, Tab+S, Tab+D, Tab+U, Tab+H,
Tab+P, Tab+M, Tab+Q, Tab+1, Tab+W, double-tap move mode (entering and
exiting via Enter, Escape and Tab), and the hudState round trip to the
preferences API.
